@@ -3,47 +3,45 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "cors-anywhere": "^0.4.3",
-    "dotenv": "^8.2.0",
-    "jss-plugin-template": "^10.1.1"
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/material": "^5.11.16",
+    "@mui/styles": "^5.11.16",
+    "date-fns": "^2.29.3",
+    "date-fns-tz": "^2.0.0",
+    "google-map-react": "^2.2.0",
+    "gtfs-realtime-bindings": "^1.1.1",
+    "lodash": "^4.17.21",
+    "prettier": "^2.8.7",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.10.0",
+    "use-debounce": "^9.0.4"
   },
   "devDependencies": {
-    "@material-ui/core": "^4.9.13",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.43",
-    "@types/google-map-react": "^1.1.7",
-    "@types/lodash": "^4.14.150",
-    "@types/moment-timezone": "^0.5.13",
-    "@types/node": "^13.13.5",
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.7",
-    "@types/react-select": "^3.0.12",
-    "auto-bind": "^4.0.0",
-    "google-map-react": "^1.1.7",
-    "gtfs-realtime-bindings": "^0.0.5",
-    "lodash": "^4.17.21",
-    "moment": "^2.25.3",
-    "moment-timezone": "^0.5.28",
-    "node-sass": "^4.14.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-countdown-circle-timer": "^2.1.2",
-    "react-dom": "^16.13.1",
-    "react-live-clock": "^4.0.5",
-    "react-scripts": "^3.4.1",
-    "react-select": "^3.1.0",
-    "typescript": "^3.8.3"
+    "@types/google-map-react": "^2.1.7",
+    "@types/lodash": "^4.14.192",
+    "@types/node": "^18.15.11",
+    "@types/react": "^17.0.56",
+    "@types/react-dom": "^17.0.19",
+    "@vitejs/plugin-react": "^3.1.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-unused-imports": "^2.0.0",
+    "rollup-plugin-visualizer": "^5.9.0",
+    "sass": "^1.61.0",
+    "typescript": "^4.9.5",
+    "vite": "^4.2.1",
+    "vite-plugin-svgr": "^2.4.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "yarn-prod": "yarn install --production",
-    "proxy": "node scripts/proxy.js"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
+    "start": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview",
+    "lint": "eslint --ext .ts,.tsx src",
+    "lint:fix": "npm run lint -- --fix",
+    "typecheck": "tsc --noEmit",
+    "tunnel": "ngrok http 3000"
   },
   "browserslist": {
     "production": [
