@@ -1,285 +1,252 @@
-import {Typography} from "@material-ui/core";
-import AppBar from "@material-ui/core/AppBar";
-import IconButton from "@material-ui/core/IconButton";
-import Snackbar from "@material-ui/core/Snackbar";
-import Toolbar from "@material-ui/core/Toolbar";
-import ExitIcon from "@material-ui/icons/ExitToApp";
-import MenuIcon from "@material-ui/icons/Menu";
-import Alert from "@material-ui/lab/Alert"
-import _ from "lodash";
-import createMoment, { TIMEZONE } from "./classes/Moment";
-import React from 'react';
-import Clock from "react-live-clock";
+import React, { useEffect, useState } from "react";
+
+import MenuIcon from "@mui/icons-material/Menu";
+import Alert from "@mui/material/Alert";
+import Snackbar from "@mui/material/Snackbar";
+import { styled } from "@mui/material/styles";
+import cloneDeep from "lodash/cloneDeep";
+import get from "lodash/get";
+import set from "lodash/set";
+import { Route, Routes, useNavigate } from "react-router-dom";
+
 import APIClient from "./classes/APIClient";
 import SettingsSet from "./classes/SettingsSet";
-import AutoBoundComponent from "./components/AutoBoundComponent";
+import CardMessage from "./components/CardMessage";
+import Clock from "./components/Clock";
+import MainAppBar from "./components/MainAppBar";
 import RefreshTimer from "./components/RefreshTimer";
-import SettingsScreen from './components/SettingsScreen';
+import { OnUpdateFunc, SettingsPane } from "./components/SettingsPane/SettingsPane";
+import SettingsScreen from "./components/SettingsScreen";
+import TripBoard from "./components/TripBoard";
 import RemindersWidget from "./components/Widget/RemindersWidget";
 import TrainMap from "./components/Widget/TrainMap";
-import TripBoard from "./components/TripBoard";
-import {ParsedVehiclePositionEntity} from "./models/GTFS/VehiclePositions";
-import {TripRequestResponseJourney} from "./models/TripPlanner/tripRequestResponseJourney";
-import CardMessage from "./components/CardMessage";
-import {styled} from "@material-ui/core/styles";
+import { ParsedVehiclePositionEntity } from "./models/GTFS/VehiclePositions";
+import { TripRequestResponseJourney } from "./models/TripPlanner/tripRequestResponseJourney";
+import { formatMediumTime } from "./util/date";
 
-interface AppState {
-    settings: SettingsSet
-    settingsMenuOpen: boolean,
-    trips: TripRequestResponseJourney[]
-    realtimeTripData: ParsedVehiclePositionEntity[],
-    isTripsRefreshing: boolean
-    lastRefreshTime: number;
-    lastApiError: string;
-}
+export default function App() {
+    const tripsInterval = 30;
+    const renderTripsInterval = 5;
+    const navigate = useNavigate();
+    const [tripsTimeoutKey, setTripsTimeoutKey] = useState(-1);
+    const [settings, setSettings] = useState(() => SettingsSet.readSettings());
+    const [hasInitialized, setHasInitialized] = useState(false);
+    const [trips, setTrips] = useState<TripRequestResponseJourney[]>([]);
+    const [realtimeTripData, setRealtimeTripData] = useState<ParsedVehiclePositionEntity[]>([]);
+    const [isTripsRefreshing, setIsTripsRefreshing] = useState(false);
+    const [lastRefreshTime, setLastRefreshTime] = useState(0);
+    const [lastApiError, setLastApiError] = useState("");
 
-export default class App extends AutoBoundComponent<{}, AppState> {
-    protected getTripsInterval = 30;
-    protected getTripsTimeoutKey;
-    protected renderTripsInterval = 5;
+    useEffect(() => {
+        (async () => {
+            await maybeLoadRemoteSettings();
+            await getTrips();
+        })();
+        return () => {
+            clearTimeout(tripsTimeoutKey);
+        };
+    }, []);
 
-    constructor(props) {
-        super(props);
-        this.state = {
-            settings: SettingsSet.readSettings(),
-            settingsMenuOpen: false,
-            trips: [],
-            realtimeTripData: [],
-            isTripsRefreshing: false,
-            lastRefreshTime: 0,
-            lastApiError: ''
-        }
-    }
-
-    componentDidMount(): void {
-        this.maybeLoadRemoteSettings();
-        this.getTrips();
-    }
-
-    componentWillUnmount(): void {
-        clearTimeout(this.getTripsTimeoutKey);
-    }
-
-    async maybeLoadRemoteSettings(): Promise<void> {
-        const { settings } = this.state;
+    const maybeLoadRemoteSettings = async () => {
         const url = new URL(window.location.href);
-        const loadUrl = url.searchParams.get('loadSettings');
-
+        const loadUrl = url.searchParams.get("loadSettings");
         if (loadUrl) {
             try {
                 const newSettings = await SettingsSet.loadRemoteSettings(loadUrl, settings);
-                this.setState({
-                    settings: newSettings
-                })
+                setSettings(newSettings);
             } catch (e) {
                 console.error(e);
             }
         }
-    }
+    };
 
-    readSettings(): void {
-        this.setState({
-            settings: SettingsSet.readSettings()
-        })
-    }
+    const readSettings = () => {
+        setSettings(SettingsSet.readSettings());
+    };
 
-    writeSettings(): void {
-        SettingsSet.writeSettings(this.state.settings);
-    }
-
-    onUpdateSetting(key: string, value: any) {
-        this.setState(prevState => {
-            const settings = _.cloneDeep(prevState.settings);
-            _.set(settings, key, value);
-            return {settings};
-        }, () => {
-            this.writeSettings();
-            if (["fromStop", "toStop"].includes(key)) {
-                this.getTrips();
+    const onUpdateSetting: OnUpdateFunc = (key, value, delta) => {
+        setSettings((prevState) => {
+            const settings = cloneDeep(prevState);
+            const currentValue = get(settings, key);
+            if (delta && typeof currentValue === "number") {
+                set(settings, key, currentValue + value);
+            } else {
+                set(settings, key, value);
             }
+
+            console.log(settings);
+
+            SettingsSet.writeSettings(settings);
+            if (["fromStop", "toStop"].includes(key)) {
+                getTrips(settings);
+            }
+
+            return settings;
         });
-    }
+    };
 
-    onResetSettings() {
+    const onResetSettings = () => {
         SettingsSet.resetSettings();
-        this.readSettings();
-    }
+        readSettings();
+    };
 
-    toggleMenu() {
-        this.setState(prevState => ({settingsMenuOpen: !prevState.settingsMenuOpen}));
-    }
+    const openMenu = () => {
+        navigate({ pathname: `/settings/${SettingsPane.GENERAL}` });
+    };
 
-    getCurrentTripLabel() {
-        const trip = this.state.settings.getConfiguredTrip();
-        if (!trip) {
-            return '';
-        }
-
-        return `: ${trip.from.disassembledName} ➡ ${trip.to.disassembledName}`
-    }
-
-    hasValidTrips(): boolean {
-        return this.state.settings.isConfiguredTrip() && this.state.trips.length > 0;
-    }
-
-    async getTrips() {
-        const {settings} = this.state;
+    const getCurrentTripLabel = () => {
         const trip = settings.getConfiguredTrip();
+        if (!trip) {
+            return "";
+        }
+        return `: ${trip.from.disassembledName} ➡ ${trip.to.disassembledName}`;
+    };
+
+    const getTrips = async (useSettings: SettingsSet = settings) => {
+        const trip = useSettings.getConfiguredTrip();
         if (!trip) {
             return;
         }
-        const {from, to} = trip;
+        const { from, to } = trip;
 
-        clearTimeout(this.getTripsTimeoutKey);
+        clearTimeout(tripsTimeoutKey);
 
-        this.setState({isTripsRefreshing: true});
-        const client = new APIClient(settings.apiKey, settings.proxyServer);
+        setIsTripsRefreshing(true);
+        const client = new APIClient();
         try {
-            const response = await client.getTrips(from, to, settings.tripCount + 1);
-            const getRealtime = settings.maps.enabled;
+            const response = await client.getTrips(from, to, useSettings.tripCount + 1);
+            const getRealtime = useSettings.maps.enabled;
 
-            this.setState({
-                trips: response.journeys || [],
-                isTripsRefreshing: getRealtime,
-                lastRefreshTime: Date.now(),
-                lastApiError: ''
-            });
+            setHasInitialized(true);
+            setTrips(response.journeys || []);
+            setIsTripsRefreshing(getRealtime);
+            setLastRefreshTime(Date.now());
+            setLastApiError("");
 
             // Now get realtime data
             if (getRealtime) {
-                const tripIds: string[] = response.journeys
-                    .map(j => j.legs[0].transportation?.properties?.RealtimeTripId)
-                    .filter(id => !!id) as string[];
+                const tripIds = response.journeys
+                    .map((j) => j.legs[0].transportation?.properties?.RealtimeTripId)
+                    .filter((id) => !!id) as string[];
                 const positionEntities = await client.getGTFSRealtime(tripIds);
-                this.setState({
-                    realtimeTripData: positionEntities,
-                    isTripsRefreshing: false,
-                    lastApiError: ''
-                });
+                setRealtimeTripData(positionEntities);
+                setIsTripsRefreshing(false);
+                setLastApiError("");
             }
         } catch (e) {
-            let message = e.message;
+            let message = e instanceof Error ? e.message : JSON.stringify(e);
 
-            if (/Failed to fetch/i.test(e.message)) {
-                message = 'Failed to fetch data from the proxy server. Please check your proxy settings (and check the proxy server is running) and try again.'
-            } else if (/401 Unauthorized/i.test(e.message)) {
-                message = 'Failed to fetch data due to an invalid API key. Please check your TfNSW API key and try again.';
+            if (/Failed to fetch/i.test(message)) {
+                message =
+                    "Failed to fetch data from the proxy server. Please check your proxy settings (and check the proxy server is running) and try again.";
+            } else if (/401 Unauthorized/i.test(message)) {
+                message =
+                    "Failed to fetch data due to an invalid API key. Please check your TfNSW API key and try again.";
             }
 
-            this.setState({
-                isTripsRefreshing: false,
-                lastApiError: message
-            });
+            setIsTripsRefreshing(false);
+            setLastApiError(message);
             console.error(e);
         } finally {
             // No matter what, try again later
-            this.scheduleTimeout();
+            scheduleTimeout();
         }
-    }
+    };
 
-    scheduleTimeout() {
-        window.clearTimeout(this.getTripsTimeoutKey);
-        this.getTripsTimeoutKey = window.setTimeout(this.getTrips, this.getTripsInterval * 1000);
-    }
+    const scheduleTimeout = () => {
+        window.clearTimeout(tripsTimeoutKey);
+        setTripsTimeoutKey(window.setTimeout(getTrips, tripsInterval * 1000));
+    };
 
-    render() {
-        const {
-            settings,
-            trips,
-            isTripsRefreshing,
-            lastRefreshTime,
-            lastApiError,
-            realtimeTripData,
-            settingsMenuOpen,
-        } = this.state;
-
-        return (
-            <div className="App">
-                <AppBar position="static" id="main-appbar">
-                    <Toolbar>
-                        <IconButton edge="start" color="inherit" aria-label="menu" onClick={this.toggleMenu}>
-                            <MenuIcon/>
-                        </IconButton>
-                        <Typography variant={"h6"}>
-                            Train Board{this.getCurrentTripLabel()}
-                        </Typography>
-                        <IconButton color="inherit" onClick={window.close}>
-                            <ExitIcon/>
-                        </IconButton>
-                    </Toolbar>
-                </AppBar>
-                <SettingsScreen
-                    menuOpen={settingsMenuOpen}
-                    settings={settings}
-                    onUpdate={(key, value) => this.onUpdateSetting(key, value)}
-                    onReset={this.onResetSettings}
-                    onClose={() => this.setState({settingsMenuOpen: false})}
+    return (
+        <div className="App">
+            <MainAppBar openMenu={openMenu} label={getCurrentTripLabel()} />
+            <Routes>
+                <Route
+                    path="settings/*"
+                    element={
+                        <SettingsScreen
+                            menuOpen={true}
+                            settings={settings}
+                            onUpdate={onUpdateSetting}
+                            onReset={onResetSettings}
+                            onClose={() => {
+                                navigate({ pathname: "/" });
+                            }}
+                        />
+                    }
                 />
+            </Routes>
 
-                <main className={[settings.maps.enabled ? "maps-enabled" : ""].join(" ")}>  
-                
+            <main className={[settings.maps.enabled ? "maps-enabled" : ""].join(" ")}>
                 {!settings.isConfiguredTrip() && (
                     <CardMessage
                         title="Welcome"
-                        body={<>
-                            Welcome to TrainBoard! To get started, open the settings menu (<MenuIconStyled />) and configure your From and To stops.
+                        body={
+                            <>
+                                Welcome to TrainBoard! To get started, open the settings menu (
+                                <MenuIconStyled />) and configure your From and To stops.
                             </>
                         }
                     />
                 )}
-                {settings.isConfiguredTrip() && !this.state.trips.length && (
-                    <CardMessage title="No trips" body="No trips were found matching your from and to stops. Please try setting different stops." />
-                )}
-                    <Snackbar
-                        anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-                        open={!!lastApiError}
-                        autoHideDuration={5000}>
-                        <Alert severity={'error'} elevation={6} variant={'filled'}>{lastApiError}</Alert>
-                    </Snackbar>
-                    <TrainMap
-                        settings={settings}
-                        trips={trips}
-                        realtimeTripData={realtimeTripData.slice(0, 2)}
+                {settings.isConfiguredTrip() && hasInitialized && trips.length === 0 && (
+                    <CardMessage
+                        title="No trips"
+                        body="No trips were found matching your from and to stops. Please try setting different stops."
                     />
-                    <div className='main-grid'>
+                )}
+                <Snackbar
+                    anchorOrigin={{ vertical: "top", horizontal: "center" }}
+                    open={!!lastApiError}
+                    autoHideDuration={5000}
+                >
+                    <Alert severity={"error"} elevation={6} variant={"filled"}>
+                        {lastApiError}
+                    </Alert>
+                </Snackbar>
+                <TrainMap
+                    settings={settings}
+                    trips={trips}
+                    realtimeTripData={realtimeTripData.slice(0, 2)}
+                />
+                <div className="main-grid">
                     <RemindersWidget settings={settings} />
                     <div className="main-wrap">
-                        <div id="trip-board-container" hidden={settings.developer.mapDebug}>
-                            {this.hasValidTrips() &&
-                            <div id="trip-board-toolbar">
-                                <Clock
-                                    format={'hh:mm:ssa'}
-                                    ticking={true}
-                                    timezone={TIMEZONE}
-                                />
-                                <div id="trip-board-timer-container">
-                                    {!!lastRefreshTime &&
-                                    <div className="status-last-refresh">Last
-                                        refreshed: {createMoment(lastRefreshTime).format("hh:mm:ssa")}</div>
-                                    }
-                                    <RefreshTimer
-                                        isRefreshing={isTripsRefreshing}
-                                        durationSeconds={this.getTripsInterval}
-                                        resetKey={lastRefreshTime}
-                                    />
+                        <div id="trip-board-container">
+                            {settings.isConfiguredTrip() && (
+                                <div id="trip-board-toolbar">
+                                    <Clock />
+                                    <div id="trip-board-timer-container">
+                                        {!!lastRefreshTime && (
+                                            <div className="status-last-refresh">
+                                                Last refreshed:{" "}
+                                                {formatMediumTime(new Date(lastRefreshTime))}
+                                            </div>
+                                        )}
+                                        <RefreshTimer
+                                            isRefreshing={isTripsRefreshing}
+                                            durationSeconds={tripsInterval}
+                                            key={lastRefreshTime}
+                                        />
+                                    </div>
                                 </div>
-                            </div>}
+                            )}
                             <TripBoard
                                 trips={trips}
                                 realtimeTripData={realtimeTripData}
                                 settings={settings}
-                                renderInterval={this.renderTripsInterval}
+                                renderInterval={renderTripsInterval}
                             />
                         </div>
                     </div>
-                    </div>
-                </main>
-            </div>
-        );
-    }
+                </div>
+            </main>
+        </div>
+    );
 }
 
 const MenuIconStyled = styled(MenuIcon)({
-    verticalAlign: 'middle',
-    lineHeight: 'initial'
+    verticalAlign: "middle",
+    lineHeight: "initial",
 });
