@@ -1,93 +1,101 @@
-import _ from "lodash";
-import React, {ChangeEvent} from "react";
-import Select, {ActionMeta, InputActionMeta, OptionsType, ValueType} from "react-select";
+import React, { useState } from "react";
+
+import { Box, CircularProgress, NativeSelect } from "@mui/material";
+import TextField from "@mui/material/TextField";
+import { useDebouncedCallback } from "use-debounce";
+
 import APIClient from "../classes/APIClient";
 import SettingsSet from "../classes/SettingsSet";
-import {StopFinderLocation} from "../models/TripPlanner/stopFinderLocation";
-import AutoBoundComponent from "./AutoBoundComponent";
+import { StopFinderLocation } from "../models/TripPlanner/stopFinderLocation";
 
 type Option = {
     label: string;
     value: string;
-}
-
-class StopSearchState {
-    query: string = "";
-    stops: StopFinderLocation[] = [];
-    options: OptionsType<Option> = [];
-}
-
+};
 interface StopSearchProps {
+    tabIndex: number;
     settings: SettingsSet;
     label: string;
-    inputId: string;
     onSelect: (stop?: StopFinderLocation) => any;
     value?: StopFinderLocation;
 }
 
-export default class StopSearch extends AutoBoundComponent<StopSearchProps, StopSearchState> {
-    protected static _uid = 0;
+export default function StopSearch(props: StopSearchProps) {
+    const { value, label } = props;
+    const [isLoading, setIsLoading] = useState(false);
+    const [query, setQuery] = useState("");
+    const [stops, setStops] = useState<StopFinderLocation[] | null>(value ? [value] : null);
+    const options =
+        stops === null
+            ? [
+                  {
+                      label: `Search for ${label}`,
+                      value: "",
+                  },
+              ]
+            : [
+                  { label: `Select location (${stops.length})`, value: "" },
+                  ...stops.map(locToOption),
+              ];
 
-    id = "";
-    debounceGetStops;
+    const debounceGetStops = useDebouncedCallback(getStops, 250);
 
-    constructor(props) {
-        super(props);
-
-        this.id = `stop-search-${StopSearch._uid++}`;
-        this.debounceGetStops = _.debounce(this.getStops, 500);
-        this.state = new StopSearchState();
+    function locToOption(loc: StopFinderLocation): Option {
+        return { label: loc.name as string, value: loc.id as string };
     }
 
-    static convertLocToOption(loc: StopFinderLocation): Option {
-        return {label: loc.name as string, value: loc.id as string};
+    function onChange(query: string) {
+        setQuery(query);
+        debounceGetStops(query);
     }
 
-    onChange(e: ChangeEvent<HTMLInputElement>) {
-        const query = e.target.value;
-
-        this.setState({
-            query: query
-        }, this.debounceGetStops);
-        
+    function onSelectInputChange(query: string) {
+        debounceGetStops(query);
     }
 
-    onSelectInputChange(query: string, actionMeta: InputActionMeta) {
-        if (actionMeta.action === 'input-change') {
-            this.debounceGetStops(query);
-        }        
+    function onSelect(selectedId: string) {
+        if (!stops) {
+            return;
+        }
+        props.onSelect(selectedId ? stops.find((stop) => stop.id === selectedId) : undefined);
     }
 
-    onSelect(selectedOption: ValueType<Option>, actionMeta: ActionMeta<Option>) {
-        //if (actionMeta.action === "select-option" || actionMeta.action === "deselect-option")
-        this.props.onSelect(selectedOption ? this.state.stops.find(stop => stop.id === (selectedOption as Option).value) : undefined);
+    async function getStops(query: string) {
+        const client = new APIClient();
+        setIsLoading(true);
+        const locations = await client.getTrainStops(query);
+
+        setIsLoading(false);
+        setStops(locations);
     }
 
-    getStops(query: string) {
-        const client = new APIClient(this.props.settings.apiKey, this.props.settings.proxyServer);
-        client.getTrainStops(query).then((locations) => {
-            const options: OptionsType<Option> = locations.map(StopSearch.convertLocToOption);
-
-            this.setState({
-                stops: locations,
-                options: options
-            });
-        }).catch((error) => {
-            console.error(error.message);
-        });
-    }
-
-    render() {
-        return (
-            <div className="stop-search">
-                <Select
-                    inputId={this.props.inputId}
-                    value={this.props.value ? StopSearch.convertLocToOption(this.props.value) : undefined}
-                    onChange={(selected, action) => this.onSelect(selected, action)}
-                    onInputChange={this.onSelectInputChange}
-                    options={this.state.options}
-                />
-            </div>
-        );
-    }
+    return (
+        <Box sx={{ display: "flex", alignItems: "center", gap: 2 }}>
+            <TextField
+                inputProps={{
+                    className: "dpad-focusable",
+                    tabIndex: 0,
+                }}
+                variant="outlined"
+                onChange={(e) => onChange(e.target.value)}
+                value={query}
+            />
+            {isLoading && <CircularProgress />}
+            <NativeSelect
+                inputProps={{
+                    className: "dpad-focusable",
+                    tabIndex: 0,
+                }}
+                value={props.value?.id}
+                onChange={(e) => onSelect(e.target.value)}
+                variant="outlined"
+            >
+                {options.map((option) => (
+                    <option key={option.value} value={option.value}>
+                        {option.label}
+                    </option>
+                ))}
+            </NativeSelect>
+        </Box>
+    );
 }
